+++ crates/ark-lsp-core/src/lsp/completions/sources/composite/call.rs
@@ -14,7 +14,7 @@ use harp::utils::r_is_function;
 use tower_lsp::lsp_types::CompletionItem;
 use tree_sitter::Node;
 
-use crate::console;
+use crate::attached;
 use crate::lsp::completions::completion_context::CompletionContext;
 use crate::lsp::completions::completion_item::completion_item_from_parameter;
 use crate::lsp::completions::sources::utils::call_node_position_type;
@@ -126,7 +126,7 @@ fn get_first_argument(context: &DocumentContext, node: &Node) -> anyhow::Result<
 
     let options = RParseEvalOptions {
         forbid_function_calls: true,
-        env: console::selected_env(),
+        env: attached::selected_env(),
     };
 
     // Try to evaluate the first argument
@@ -190,7 +190,7 @@ fn completions_from_session_arguments(
     // for S3 methods based on `object`.
     let r_callable = harp::parse_eval(callable, RParseEvalOptions {
         forbid_function_calls: true,
-        env: console::selected_env(),
+        env: attached::selected_env(),
     });
 
     let r_callable = match r_callable {